fixed bugs with voice login and bt login overlap

+++ main.py
@@ -5,6 +5,7 @@ import bluetooth
 import time
 
 def is_user(ermrest):
+	#checks if a user is logged in at the moment
 	try:
         	data = ermrest.get_data(7,"session_info")[0]
 	except:
@@ -15,18 +16,20 @@ def is_user(ermrest):
         return False
 
 def action(phone,ermrest):
+	#attempts to log user in and start a Jarvis session
         try:
                 users = ermrest.get_data(8,"users")
-		user_info = None
 		for user in users:
 			if user['phone_identification'] in phone:
-				user_info = user
+				username = user['username']
+				break
 
-                data = {"user":user_info['username'],"jarvis_response":None,"current_experiment_id":None}
+                data = {"user":username,"jarvis_response":None,"current_experiment_id":None}
                 try:
                         ermrest.delete_data(7,"session_info")
                 except:
                         print("No data in session info")
+
                 ermrest.put_data(7,"session_info",data)
                 return True
                 
@@ -34,6 +37,7 @@ def action(phone,ermrest):
                 return False
 
 def check_user_exists(ermrest):
+	#checks if the user is still in the area
 	devices = bluetooth.discover_devices()
 	device_names = []
 	for device in devices:
@@ -41,12 +45,11 @@ def check_user_exists(ermrest):
 
 	current_user = ermrest.get_data(7,"session_info")[0]['user']	
 	try:
-		user_info = ermrest.get_data(8,"users","/username="+current_user)[0]
+		phone = ermrest.get_data(8,"users","/username="+current_user)[0]['phone_identification']
 	except:
-		user_info = {'phone_identification':"Empty_id_slot"} #A string because lookup_name can return None
-	phone = user_info['phone_identification']
+		phone = 'NULL' #A string because lookup_name can return None
 
-	if phone in devices or phone in device_names:
+	if phone in devices or phone in device_names: #checks both because mac/name are boot valid forms of identification
 		return True
 	return False
  
@@ -57,31 +60,38 @@ def main():
 	phone_retriever = Phone_retriever()
 	logger = phone_retriever.logger
 	ermrest = ErmrestHandler("ec2-54-172-182-170.compute-1.amazonaws.com","root","root") 
+	users = ermrest.get_data(8,"users")[0]
 	fail_counter = 0
 	timer = time.time()
 	reset_timer = time.time()
 	bootup_run = True
 	logged_in = False 
+	voice_login = False
 	user = "None"
-	run_interval = 8
+	run_interval = 8 #run_interval changes because it takes about 8 seconds after rc.local is run to complete bootup sequence of raspberry pi
 
 	while True: #main loop
 		if (time.time()-timer > run_interval): #checks when to run so it isn't constantly running
 			devices = bluetooth.discover_devices()
-			users = ermrest.get_data(8,"users")[0]
 			#checks if user is logged in AND if the user is still in the area.
 			if (is_user(ermrest)):
 				user = ermrest.get_data(7,"session_info")[0]['user']
-				if (logged_in == False):
+				if (logged_in == False): #logs log in even if user logged in through voice command not bluetooth
 					print("User {} log in at: ".format(user)+time.asctime(time.localtime(time.time()))) 
 					print >> logger, "User {} log in at: ".format(user)+time.asctime(time.localtime(time.time()))
 					logged_in = True
+					voice_login = True
+					continue
 				
-				if (check_user_exists(ermrest)):
+				if (check_user_exists(ermrest) and voice_login != True): 
+					#must make sure that user didn't log in with voice
+					#so that the user isn't logged out because they didn't use bluetooth 
 					timer = time.time()
 					fail_counter = 0
 					continue
-				else: #leeway for some signal drops
+
+				elif (voice_login != True): #leeway for some signal drops
+					timer = time.time()
 					if (fail_counter >= 2):
 						ermrest.delete_data(7,"session_info")
 						ermrest.put_data(7,"session_info",empty_table)
@@ -92,26 +102,29 @@ def main():
 					else:
 						fail_counter += 1
 						continue
-					timer = time.time()
-			elif (logged_in):
+			elif (logged_in): #if no user is in session_info and logged_in is still true log out user
 				print("User {} log out at: ".format(user)+time.asctime(time.localtime(time.time())))
 				print >> logger,"User {} log out at: ".format(user)+time.asctime(time.localtime(time.time()))
 				logged_in = False
+				voice_login = False
 						
 
-			timer = time.time()
-			nearest_phone = phone_retriever.get_nearest_phone()
+			timer = time.time() #always reset timer
 
-			if(action(nearest_phone,ermrest)):
-				print("User {} log in at: ".format(user)+time.asctime(time.localtime(time.time()))) 
-				print >> logger, "User {} log in at: ".format(user)+time.asctime(time.localtime(time.time()))
-				logged_in = True
-				try:
-					ermrest.delete_data(7,"step_completed")
-				except:
-					pass
-				ermrest.put_data(7,"step_completed",{"completed_step":None})
+			if (voice_login == False):
+				nearest_phone = phone_retriever.get_nearest_phone()
+				if(action(nearest_phone,ermrest)): #if user is successfully logged in
+					print("User {} log in at: ".format(user)+time.asctime(time.localtime(time.time()))) 
+					print >> logger, "User {} log in at: ".format(user)+time.asctime(time.localtime(time.time()))
+					logged_in = True
+					#must reset everything the past user might have left behind
+					try:
+						ermrest.delete_data(7,"step_completed")
+					except:
+						pass
+					ermrest.put_data(7,"step_completed",{"completed_step":None})
 			if (bootup_run):
+				#change the run_interval after bootup_completes
 				run_interval = 5
 				bootup_run = False
